Move client files to utils folder
